moved itemtooltip code to separate file due to random crashes on startup
